split EncryptedPool from lib.rs add test template for EncryptedPool
